+++ app/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/di/FragmentModule.kt
@@ -2,9 +2,9 @@ package jp.chau2chaun2.honkot.sample.multimodule.di
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
-import jp.chau2chaun2.honkot.sample.multimodule.ui.ListQiitaDocFragment
-import jp.chau2chaun2.honkot.sample.multimodule.ui.ListDatabaseFragment
-import jp.chau2chaun2.honkot.sample.multimodule.ui.SelectFragment
+import jp.chau2chaun2.honkot.sample.multimodule.ui.fragment.ListQiitaDocFragment
+import jp.chau2chaun2.honkot.sample.multimodule.ui.fragment.ListDatabaseFragment
+import jp.chau2chaun2.honkot.sample.multimodule.ui.fragment.SelectFragment
 
 @Module
 abstract class FragmentModule {
