+++ app/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/ui/adapter/ListDatabaseAdapter.kt
@@ -1,4 +1,4 @@
-package jp.chau2chaun2.honkot.sample.multimodule.ui
+package jp.chau2chaun2.honkot.sample.multimodule.ui.adapter
 
 import android.content.res.Resources
 import android.view.LayoutInflater
@@ -10,6 +10,7 @@ import androidx.lifecycle.LifecycleOwner
 import jp.chau2chaun2.honkot.sample.multimodule.R
 import jp.chau2chaun2.honkot.sample.multimodule.databinding.RowListDatabaseBinding
 import jp.chau2chaun2.honkot.sample.multimodule.model.Repo
+import jp.chau2chaun2.honkot.sample.multimodule.vm.RowListDatabaseViewModel
 
 class ListDatabaseAdapter(
     private val viewLifecycleOwner: LifecycleOwner,
@@ -31,7 +32,11 @@ class ListDatabaseAdapter(
         } ?: run {
             DataBindingUtil.inflate<RowListDatabaseBinding>(layoutInflater, R.layout.row_list_database, null, false).also {
                 it.lifecycleOwner = viewLifecycleOwner
-                it.viewModel = RowListDatabaseViewModel(resources, item)
+                it.viewModel =
+                    RowListDatabaseViewModel(
+                        resources,
+                        item
+                    )
             }
         }
         binding.executePendingBindings()
