+++ app/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/vm/RowListQiitaDocViewModel.kt
@@ -1,4 +1,4 @@
-package jp.chau2chaun2.honkot.sample.multimodule.ui
+package jp.chau2chaun2.honkot.sample.multimodule.vm
 
 import android.graphics.Bitmap
 import androidx.lifecycle.*
