+++ app/src/main/res/layout/row_list_database.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="jp.chau2chaun2.honkot.sample.multimodule.ui.RowListDatabaseViewModel" />
+            type="jp.chau2chaun2.honkot.sample.multimodule.vm.RowListDatabaseViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
