+++ app/src/main/res/layout/row_list_qiita_doc.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="jp.chau2chaun2.honkot.sample.multimodule.ui.RowListQiitaDocViewModel" />
+            type="jp.chau2chaun2.honkot.sample.multimodule.vm.RowListQiitaDocViewModel" />
 
         <import type="android.view.View" />
     </data>
